Exception handling - update e delete atualizacao

+++ src/main/java/com/requests/project/services/UserService.java
@@ -1,10 +1,12 @@
 package com.requests.project.services;
 import java.util.List;
 import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
+
 import com.requests.project.entities.User;
 import com.requests.project.repositories.UserRepository;
 import com.requests.project.services.exceptions.DatabaseException;
@@ -17,9 +19,11 @@ public class UserService {
 
 	@Autowired
 	private UserRepository repository;
+	
 	public List<User> findAll() {
 		return repository.findAll();
 	}
+	
 	public User findById(Long id) {
 		Optional<User> obj = repository.findById(id);
 		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
@@ -30,14 +34,16 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		try {
-			repository.deleteById(id);
-		} catch (EmptyResultDataAccessException e) {
-			throw new ResourceNotFoundException(id);
-		} catch (DataIntegrityViolationException e) {
-			throw new DatabaseException(e.getMessage());
-		}
-	}
+	    try {
+	        if (repository.existsById(id)) {
+	            repository.deleteById(id);			
+	        } else {				
+	            throw new ResourceNotFoundException(id);			
+	        }		
+	    } catch (DataIntegrityViolationException e) {			
+	        throw new DatabaseException(e.getMessage());		
+	    }	
+	} 
 
 	public User update(Long id, User obj) {
 		try {
